backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
google-generativeai
python-dotenv
google-genai